% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_assessments.R
\name{analyze_assessments}
\alias{analyze_assessments}
\title{Title}
\usage{
analyze_assessments(
  file = system.file("assessments/assessments.yaml", package = "hoaTools")
)
}
\arguments{
\item{file}{}
}
\description{
Title
}
